Migrating to Flask
So I have spent my time not working on the github pages site learning HTML and a bit of CSS. After doing that I picked up Flask and fired up a new version of the site. It's kind of back to basics again, but now I can start implementing some fancier stuff.